09022023-Final

+++ Data/Entities/Course.cs
@@ -7,6 +7,6 @@
         // Definir as Props
         public int Id { get; set; }
         public string NomeCurso { get; set; }
-        public float Mensalidade { get; set; }
+        public double Mensalidade { get; set; }
     }
 }
